Major overhaul of Blender import/export space conversion
To implement reliable animation export of rotations, Blender will now apply a corrective rotation to any root node on import & export rather than try to bake this additional rotation into every node transform. So on import we get root nodes with +90 X rotation and this is reversed (in code, not in the scene) on export.