Adds support for Blender 4.0.0
Blender 4.0.0 has changed the structure of the principled BSDF node such
that we have more direct control over the Specular functions. This
requires calls to the original Specular node to be renamed to fit the
new pattern.

This pull request includes fallback for older versions of Blender to
maintain compatibility. Tested on 3.6.4 and 4.0.0 beta.

Note: If you load a 3.6.4 material into Blender 4.0 with an actively used
Specular node, the conversion to the new node will add a Mix node for the
diffuse color for Specular Tint. In the case for materials created via
io_pdx_mesh import, this mix node is left blank and is unneeded for
native import of meshes.